docshell/test/mochitest/test_javascript_sandboxed_popup.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>

<body>
<iframe srcdoc="<a href='javascript:opener.parent.ok(false, `The JS ran!`)' target=_blank rel=opener>click</a>"
  sandbox="allow-popups allow-same-origin"></iframe>

<script>
add_task(async function() {
  let promise = new Promise(resolve =>{
    SpecialPowers.addObserver(function obs(subject) {
      is(subject.opener, window[0],
         "blocked javascript URI should have been targeting the pop-up document");
      subject.close();
      SpecialPowers.removeObserver(obs, "javascript-uri-blocked-by-sandbox");
      resolve();
    }, "javascript-uri-blocked-by-sandbox");
  });
  document.querySelector("iframe").contentDocument.querySelector("a").click();
  await promise;
});
</script>
</body>
